Temporarily implement IsChatOpen
The logic ` if (IsKeyTapped(Keys.Escape) && Main.drawingPlayerChat)`
Was never triggered, because Terraria already updated the `drawingPlayerChat` boolean to false. Would need to have a PreUpdate hook for this I'm guessing.
For now added our own bool to track the state of the chat open, to clear out the chat if Escape is pressed while chat is open.