Implement SendBytes hook
This is new to me so I have a lot to learn. I was able to inject the hook and push the variables onto the InvokeSendBytes call.
I could not figure out yet how to inject an if statement and return if the event was handled.
Also did not inject the SentBytes invoke yet.